Changes are made to the data structure and value types related to the start date of tracking in the tracker object.
Changes in initialValues:

startDate was previously a string representing the date in string format (toDateString()), but it is now an object of type Date, which improves type safety and provides a more accurate representation of the date.
Changes in the onSubmit type:

The startDate parameter in the onSubmit function was previously of type string, but it is now of type Date, reflecting the change in the tracker structure.
Changes in the TTrackerUpdate type:

The TTrackerUpdate type has also been updated to reflect the change in the startDate type from string to Date.
These changes enhance type safety and the accuracy of data related to the start date of tracking in the project's code.